Moved function that deal with printing return
